Logins

+++ cal/views.py
@@ -1,15 +1,35 @@
-from datetime import datetime, timedelta, date
+from datetime import datetime, timedelta, date, timezone
 import calendar
 from django.shortcuts import render, get_object_or_404
 from django.http import HttpResponse, HttpResponseRedirect
 from django.views import generic
 from django.utils.safestring import mark_safe
+from django.core.exceptions import PermissionDenied
 from .models import *
 from .utils import MonthlyCalendar
 from .forms import EventForm
+from canvasapi import Canvas
+from .local_settings import *
+import pytz
+import tzlocal
 
 class LoginView(generic.TemplateView):
     template_name = 'cal/login.html'
+
+class CanvasItemListView(generic.ListView):
+    model = Event
+    context_object_name = 'event_list'
+    template_name = 'cal/canvas_item_list.html'
+
+    def get_context_data(self, *args, **kwargs):
+        context = super().get_context_data(**kwargs)
+        user, courses, assignments = access_canvas()
+        asgn_events = []
+        for a in assignments:
+            asgn_events.append(dict_to_event(a, user=self.request.user))
+            asgn_events[-1].save()
+        context['event_list'] = asgn_events
+        return context
 
 class DailyCalendarView(generic.ListView):
     model = Event
@@ -19,7 +39,8 @@ class DailyCalendarView(generic.ListView):
     def get_queryset(self):
         self.datestring = self.request.GET.get('date', None)
         self.day = date.fromisoformat(self.datestring)
-        return Event.objects.filter(start_time__contains=self.day)
+        qs = Event.objects.filter(user=self.request.user, start_date__contains=self.day)
+        return qs
 
 class MonthlyCalendarView(generic.ListView):
     model = Event
@@ -27,8 +48,10 @@ class MonthlyCalendarView(generic.ListView):
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
+        if not self.request.user.is_authenticated:
+            return None
         d = get_date(self.request.GET.get('month', None))
-        cal = MonthlyCalendar(d.year, d.month)
+        cal = MonthlyCalendar(self.request, d.year, d.month)
         html_cal = cal.format()
         context['monthlycalendar'] = mark_safe(html_cal)
         context['prev_month'] = prev_month(d)
@@ -64,7 +87,10 @@ def event(request, event_id=None):
     else:   # otherwise, use a blank Event
         instance = Event()
     
-    form = EventForm(request.POST or None, instance=instance)   # create a form to create or edit the event
+    if instance.user != request.user:
+        raise PermissionDenied
+
+    form = EventForm(request.POST or None, instance=instance, user=request.user)   # create a form to create or edit the event
     if request.POST and form.is_valid():
         # save the edited Event
         form.save()
@@ -74,7 +100,60 @@ def event(request, event_id=None):
 def delete_event(request, event_id=None):
     instance = get_object_or_404(Event, pk=event_id)
 
+    if request.user != instance.user:
+        raise PermissionDenied
     if request.method == 'POST':
         instance.delete()
         return HttpResponseRedirect(reverse('cal:monthlycalendar'))
     return render(request, 'cal/confirm_delete.html', {'evt_id':event_id})
+
+def access_canvas():
+    metadata = Metadata.load()
+
+    # Load the last call and localize it to UTC.
+    last_call = metadata.last_canvas_call or pytz.utc.localize(datetime.utcnow())
+    last_call = last_call.replace(tzinfo=pytz.UTC)
+
+    # Set call_time to the time right before we make the API call.
+    call_time = pytz.utc.localize(datetime.utcnow())
+
+    canvas = Canvas(API_URL, API_KEY)
+    user = canvas.get_current_user()
+    courses = user.get_courses(enrollment_status='active')
+    assignments = []
+
+    # For each of the user's courses for the current term, finds all assignments created after last_call,
+    # converts them to dictionaries, and adds them to a list.
+    for c in courses:
+        if 'enrollment_term_id' in c.__dict__.keys() and c.__dict__['enrollment_term_id']==108:
+            asgn = c.get_assignments()
+            for a in asgn:
+                a = a.__dict__
+                try:
+                    created = pytz.utc.localize(datetime.strptime(a['created_at'], '%Y-%m-%dT%H:%M:%SZ'))
+                    if datetime.strptime(a['due_at'], '%Y-%m-%dT%H:%M:%SZ') and created and created >= last_call:
+                        a['course_name'] = c.name
+                        assignments.append(a)
+                except Exception as ex:
+                    print(ex)
+    
+    metadata.last_canvas_call = call_time
+    metadata.save()
+    return user, courses, assignments
+
+def dict_to_event(assignment, user):
+    '''
+    Takes an assignment (as a dictionary) and returns an event with the same name, start_time equal to the due date,
+    and the course name as the description.
+    '''
+    dc = assignment
+    st = pytz.utc.localize(datetime.strptime(dc['due_at'], '%Y-%m-%dT%H:%M:%SZ'))
+    print(st)
+    st2 = st.astimezone(tzlocal.get_localzone())
+    ev = Event.objects.create(title=dc['name'], start_time=st2, start_date=st2.date(), start_hm=st2.time(),
+     description=dc['course_name'], user=user)
+    return ev
+
+def error_500(request, *args, **kwargs):
+    data = {}
+    return render(request, 'cal/500.html', data)
